CurrentUser

+++ sql/archive/t.branch.sql
@@ -16,10 +16,10 @@ CREATE TABLE `branch` (
     `return_store_id` INT NULL COMMENT '默认退货仓库',
     `purchase_store_id` INT NULL COMMENT '默认进货仓库',
     `memo` VARCHAR(100) NOT NULL DEFAULT '' COMMENT '备注',
-    `create_oper_id` INT NOT NULL COMMENT '创建人',
+    `create_oper_id` BIGINT NOT NULL COMMENT '创建人',
 	`create_oper_name` VARCHAR(45) NOT NULL,
 	`create_time` DATETIME NOT NULL COMMENT '创建时间',
-    `last_update_oper_id` INT NOT NULL COMMENT '最后修改人',
+    `last_update_oper_id` BIGINT NOT NULL COMMENT '最后修改人',
 	`last_update_oper_name` VARCHAR(45) NOT NULL,
 	`last_update_time` DATETIME NOT NULL COMMENT '最后修改时间',
     PRIMARY KEY (`id`))
